requirements.txt
Django==3.1.3
requests==2.25.0
beautifulsoup4==4.9.3
feedparser==6.0.2
gunicorn==20.0.4
psycopg2==2.8.6
dj_database_url==0.5.0
lxml==4.6.2
djangorestframework==3.12.2
django-filter==2.4.0
